error.stack is assumed to exist
I had a stack overflow in my code which does not generate an
error.stack. As a result I was seeing cannot call method split() on
undefined (with no stack). Took a while to trace that to here.

This doesn’t magically show the cause of the overflow, but at least it
shows the “correct” error.